+++ src/components/vertical-menu/responsive-vertical-menu/responsive-vertical-menu-divider/index.ts
@@ -1,0 +1,2 @@
+export { default as ResponsiveVerticalMenuDivider } from "./responsive-vertical-menu-divider.component";
+export type { ResponsiveVerticalMenuDividerProps } from "./responsive-vertical-menu-divider.component";